fix: remaining migration issues
