+++ package.json
@@ -69,7 +69,10 @@
   },
   "pnpm": {
     "overrides": {
-      "node-abi": "4.9.0"
+      "node-abi": "4.9.0",
+      "prosemirror-model": "1.25.2",
+      "yjs": "13.6.16",
+      "zod-validation-error": "3.4.1"
     }
   }
 }
