+++ packages/ai-workspace-common/src/modules/selection-menu/use-selection-context.ts
@@ -1,9 +1,8 @@
 import { useCallback, useEffect, useState } from 'react';
-import { IContextItem } from '@refly/common-types';
+import { IContextItem, ContextTarget } from '@refly/common-types';
 import {
   useContextPanelStoreShallow,
   useContextPanelStore,
-  ContextTarget,
   useCanvasStoreShallow,
 } from '@refly/stores';
 import { getSelectionNodesMarkdown } from '@refly/utils/html2md';
