+++ packages/utils/package.json
@@ -58,7 +58,7 @@
     "mitt": "~3.0.1",
     "node-html-parser": "^6.1.13",
     "prosemirror-markdown": "~1.13.0",
-    "prosemirror-model": "1.22.3",
+    "prosemirror-model": "1.25.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "tailwind-merge": "^2.2.2",
