+++ packages/web-core/src/components/landing-page-partials/HeroHome.tsx
@@ -9,7 +9,7 @@ import { FaGithub } from 'react-icons/fa6';
 
 import { Button, Modal } from 'antd';
 import BlurImage from '../../components/common/BlurImage';
-import { useAuthStoreShallow } from '@refly-packages/ai-workspace-common/stores/auth';
+import { useAuthStoreShallow } from '@refly/stores';
 import { useState } from 'react';
 
 function HeroHome() {
