+++ packages/web-core/src/pages/skill-response-share/index.tsx
@@ -1,6 +1,6 @@
 import { useParams } from 'react-router-dom';
 import { useTranslation } from 'react-i18next';
-import { useSiderStoreShallow } from '@refly-packages/ai-workspace-common/stores/sider';
+import { useSiderStoreShallow } from '@refly/stores';
 import { Result } from 'antd';
 import { SimpleStepCard } from '@refly-packages/ai-workspace-common/components/slideshow/components/SimpleStepCard';
 import { useFetchShareData } from '@refly-packages/ai-workspace-common/hooks/use-fetch-share-data';
